Moved Fish Requirements extraction, also overlay and Y meter initialization and drawing, from FishRequirementsRecipeCategory to FishRequirementsRecipeWrapper.  - Also, uses JEI's DrawableResource class to improve readability and (maybe) compatibility.
 With this, minigame and Y meters are independent for each fish, so when the player has Video Scale Normal or smaller and JEI shows 2 fishes on the same screen, each information shows accordingly.

+++ src/main/java/kameib/localizator/data/Texture.java
@@ -4,9 +4,9 @@ import net.minecraft.util.ResourceLocation;
 
 public class Texture {
     public ResourceLocation texture;
-    public float textureWidth, textureHeight;
+    public int textureWidth, textureHeight;
     
-    public Texture(ResourceLocation texture, float textureWidth, float textureHeight) {
+    public Texture(ResourceLocation texture, int textureWidth, int textureHeight) {
         this.texture = texture;
         this.textureWidth = textureWidth;
         this.textureHeight = textureHeight;
